move maps and types
